added check for offscreen canvas and fallback
